[FIXED JENKINS-46191] Do not cast empty declarations.
Before this change, "String foo;" would result in
```
BUG! exception in phase 'class generation' in source unit 'Script1.groovy' Error while popping argument from operand stack tracker in class Script1 method java.lang.Object run().
```

This is because in that scenario, the right hand side is an
`EmptyExpression`, which ends up breaking all the things when you try
to cast it to something else. So we shouldn't be trying to do a cast
at all here, and instead just transform the declaration expression
(which actually may not be needed itself, but better safe than sorry
to still transform the left hand side's variable expression, which is
basically all that will happen).